Precision updates
Made everything into doubles.
Found a better approximation for the robot
Is close but still unstable

+++ include/robot/AutoLQR.h
@@ -40,14 +40,14 @@ public:
      * @param A Pointer to state matrix (stateSize x stateSize)
      * @return true if successful, false otherwise
      */
-    bool setStateMatrix(const float* A);
+    bool setStateMatrix(const double* A);
 
     /**
      * @brief Set the input matrix B
      * @param B Pointer to input matrix (stateSize x controlSize)
      * @return true if successful, false otherwise
      */
-    bool setInputMatrix(const float* B);
+    bool setInputMatrix(const double* B);
 
     /**
      * @brief Set the cost matrices Q and R
@@ -55,7 +55,7 @@ public:
      * @param R Pointer to control cost matrix (controlSize x controlSize)
      * @return true if successful, false otherwise
      */
-    bool setCostMatrices(const float* Q, const float* R);
+    bool setCostMatrices(const double* Q, const double* R);
 
     /**
      * @brief Compute optimal feedback gains using current solver parameters
@@ -69,14 +69,14 @@ public:
      * @param tolerance Convergence tolerance for DARE solver
      * @return true if successful, false if computation fails
      */
-    bool computeGains(int maxIterations, float tolerance);
+    bool computeGains(int maxIterations, double tolerance);
 
     /**
      * @brief Set solver parameters for DARE computation
      * @param maxIterations Maximum number of iterations (default: 100)
      * @param tolerance Convergence tolerance (default: 1e-6)
      */
-    void setSolverParameters(int maxIterations, float tolerance);
+    void setSolverParameters(int maxIterations, double tolerance);
 
     /**
      * @brief Get current maximum iterations setting
@@ -88,25 +88,25 @@ public:
      * @brief Get current tolerance setting
      * @return Tolerance value
      */
-    float getTolerance() const;
+    double getTolerance() const;
 
     /**
      * @brief Update the controller with current state
      * @param currentState Pointer to current state vector (stateSize)
      */
-    void updateState(const float* currentState);
+    void updateState(const double* currentState);
 
     /**
      * @brief Calculate control inputs based on current state
      * @param controlOutput Pointer to control output vector (controlSize)
      */
-    void calculateControl(float* controlOutput);
+    void calculateControl(double* controlOutput);
 
     /**
      * @brief Set pre-computed gain values
      * @param K Pointer to gain matrix (controlSize x stateSize)
      */
-    void setGains(const float* K);
+    void setGains(const double* K);
 
     /**
      * @brief Check if the system is controllable using full controllability matrix
@@ -119,7 +119,7 @@ public:
      * @param C Pointer to output array (stateSize x stateSize*controlSize)
      * @return true if successful, false otherwise
      */
-    bool getControllabilityMatrix(float* C);
+    bool getControllabilityMatrix(double* C);
 
     /**
      * @brief Get the rank of the controllability matrix
@@ -143,50 +143,50 @@ public:
      * @brief Get the solution of the Riccati equation
      * @return Pointer to the P matrix (stateSize x stateSize)
      */
-    const float* getRicattiSolution() const;
+    const double* getRicattiSolution() const;
 
     /**
      * @brief Estimate feedforward gain for steady-state tracking
      * @param ffGain Pointer to feedforward gain vector (controlSize)
      * @param desiredState Pointer to desired state vector (stateSize)
      */
-    void estimateFeedforwardGain(float* ffGain, const float* desiredState);
+    void estimateFeedforwardGain(double* ffGain, const double* desiredState);
 
     /**
      * @brief Estimate time to convergence
      * @param convergenceThreshold Threshold for considering system converged (default: 0.05)
      * @return Estimated time in seconds, or -1 if estimation fails
      */
-    float estimateConvergenceTime(float convergenceThreshold = 0.05f);
+    double estimateConvergenceTime(double convergenceThreshold = 0.05f);
 
     /**
      * @brief Export computed gains to external array
      * @param exportedK Pointer to destination array (controlSize x stateSize)
      * @return true if successful, false otherwise
      */
-    bool exportGains(float* exportedK);
+    bool exportGains(double* exportedK);
 
     /**
      * @brief Calculate expected cost from current state
      * @return Expected cost value, or -1 if calculation fails
      */
-    float calculateExpectedCost();
+    double calculateExpectedCost();
 
 private:
     int stateSize; ///< Number of state variables
     int controlSize; ///< Number of control inputs
 
-    float* A; ///< State matrix
-    float* B; ///< Input matrix
-    float* Q; ///< State cost matrix
-    float* R; ///< Control cost matrix
-    float* K; ///< Control gain matrix
-    float* state; ///< Current state
-    float* P; ///< Riccati equation solution
+    double* A; ///< State matrix
+    double* B; ///< Input matrix
+    double* Q; ///< State cost matrix
+    double* R; ///< Control cost matrix
+    double* K; ///< Control gain matrix
+    double* state; ///< Current state
+    double* P; ///< Riccati equation solution
 
     // Solver parameters
     int solverMaxIterations; ///< Maximum DARE solver iterations
-    float solverTolerance; ///< DARE convergence tolerance
+    double solverTolerance; ///< DARE convergence tolerance
 
     // Helper functions
 
@@ -199,7 +199,7 @@ private:
      * @param cols1 Columns in first matrix / Rows in second matrix
      * @param cols2 Columns in second matrix
      */
-    void matrixMultiply(const float* m1, const float* m2, float* result, int rows1, int cols1, int cols2);
+    void matrixMultiply(const double* m1, const double* m2, double* result, int rows1, int cols1, int cols2);
 
     /**
      * @brief Add two matrices
@@ -209,7 +209,7 @@ private:
      * @param rows Rows in matrices
      * @param cols Columns in matrices
      */
-    void matrixAdd(const float* m1, const float* m2, float* result, int rows, int cols);
+    void matrixAdd(const double* m1, const double* m2, double* result, int rows, int cols);
 
     /**
      * @brief Subtract second matrix from first
@@ -219,7 +219,7 @@ private:
      * @param rows Rows in matrices
      * @param cols Columns in matrices
      */
-    void matrixSubtract(const float* m1, const float* m2, float* result, int rows, int cols);
+    void matrixSubtract(const double* m1, const double* m2, double* result, int rows, int cols);
 
     /**
      * @brief Transpose a matrix
@@ -228,7 +228,7 @@ private:
      * @param rows Rows in input matrix
      * @param cols Columns in input matrix
      */
-    void transposeMatrix(const float* matrix, float* result, int rows, int cols);
+    void transposeMatrix(const double* matrix, double* result, int rows, int cols);
 
     /**
      * @brief Invert a small matrix (1x1, 2x2, 3x3)
@@ -237,7 +237,7 @@ private:
      * @param n Matrix size
      * @return true if successful, false if matrix is singular
      */
-    bool invertMatrix(const float* matrix, float* result, int n);
+    bool invertMatrix(const double* matrix, double* result, int n);
 
     /**
      * @brief Compute the optimal gain matrix by solving DARE
@@ -249,7 +249,7 @@ private:
      * @brief Build the controllability matrix internally
      * @param C Pointer to output array (stateSize x stateSize*controlSize)
      */
-    void buildControllabilityMatrix(float* C);
+    void buildControllabilityMatrix(double* C);
 
     /**
      * @brief Compute rank of a matrix using Gaussian elimination
@@ -258,7 +258,7 @@ private:
      * @param cols Number of columns
      * @return Rank of the matrix
      */
-    int computeMatrixRank(float* matrix, int rows, int cols);
+    int computeMatrixRank(double* matrix, int rows, int cols);
 
     /**
      * @brief Compute determinant of a square matrix (up to 4x4)
@@ -266,7 +266,7 @@ private:
      * @param n Matrix size
      * @return Determinant value
      */
-    float computeDeterminant(const float* matrix, int n);
+    double computeDeterminant(const double* matrix, int n);
 };
 
 #endif
